# contracts/openapi/console-v1.yaml
openapi: 3.1.0
info: { title: Open Platform Console API, version: 1.0.0, license: { name: Proprietary, identifier: LicenseRef-Proprietary } }
servers: [{ url: /console/api/v1 }]
paths:
  /sessions/csrf:
    get:
      operationId: getLoginCsrfToken
      summary: 获取登录所需的 CSRF 令牌
      security: []
      responses:
        '200':
          description: CSRF token initialized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CsrfToken' }
        '403':
          description: 请求被安全策略拒绝
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /sessions:
    post:
      operationId: createConsoleSession
      summary: 使用账号和密码建立控制台会话
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: 会话已建立
          headers:
            Set-Cookie:
              description: 生产环境返回 `SESSION=<opaque>; Path=/; HttpOnly; Secure; SameSite=Strict`，并轮换登录前会话标识。
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionContext' }
        '400': { description: 请求无效, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: 凭据无效, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '423':
          description: 登录暂时锁定
          headers: { Retry-After: { schema: { type: integer, minimum: 1 } } }
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '429':
          description: 来源 IP 登录失败过多
          headers: { Retry-After: { schema: { type: integer, minimum: 1 } } }
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '503': { description: 认证服务暂时不可用, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /session:
    get:
      operationId: getConsoleSession
      summary: 获取当前控制台会话
      security: [{ ConsoleSession: [] }]
      responses:
        '200': { description: 当前会话, content: { application/json: { schema: { $ref: '#/components/schemas/SessionContext' } } } }
        '401': { description: 未认证, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '503': { description: 认证服务暂时不可用, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    delete:
      operationId: deleteConsoleSession
      summary: 终止当前控制台会话
      security: [{ ConsoleSession: [] }]
      responses:
        '204':
          description: 会话已终止
          headers:
            Set-Cookie:
              description: 清除当前 `SESSION` Cookie（空值且 Max-Age=0）。
              schema: { type: string }
        '401': { description: 未认证, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: CSRF 校验失败, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '503': { description: 会话存储暂时不可用, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /onboarding/status:
    get:
      operationId: getOnboardingStatus
      summary: 获取当前登录客户的入驻审核结果
      security: [{ ConsoleSession: [] }]
      responses:
        '200': { description: 本人入驻审核结果, content: { application/json: { schema: { $ref: '#/components/schemas/OnboardingStatusResponse' } } } }
        '401': { description: 会话无效, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '503': { description: 身份服务暂时不可用, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /applications:
    get:
      operationId: listApplications
      summary: 查询当前企业的唯一应用
      security: [{ ConsoleSession: [] }]
      responses:
        '200':
          description: 当前企业应用列表
          headers: { X-Request-ID: { required: true, schema: { type: string, minLength: 1 } } }
          content: { application/json: { schema: { type: array, maxItems: 1, items: { $ref: '#/components/schemas/ApplicationResponse' } } } }
        '401': { description: 会话无效, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: 入驻未通过, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '503': { description: 应用服务暂时不可用, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    post:
      operationId: createApplication
      summary: 创建当前企业唯一应用并首次领取沙箱密钥
      security: [{ ConsoleSession: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ApplicationRequest' } } } }
      responses:
        '201':
          description: 应用及沙箱凭证已创建，AppSecret 仅在本响应出现
          headers: { Cache-Control: { schema: { type: string, const: no-store } } }
          content: { application/json: { schema: { $ref: '#/components/schemas/ApplicationCreatedResponse' } } }
        '400': { description: 字段校验失败, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: 会话无效, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: 入驻未通过或 CSRF 失败, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: 当前企业已有应用, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '503': { description: 应用或凭证服务暂时不可用, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /applications/{applicationId}/permissions:
    parameters:
      - { name: applicationId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      operationId: listApplicationPermissions
      summary: 查询当前应用的三项接口权限
      security: [{ ConsoleSession: [] }]
      responses:
        '200':
          description: 三项固定接口权限
          headers: { X-Request-ID: { required: true, schema: { type: string, minLength: 1 } } }
          content: { application/json: { schema: { type: array, minItems: 3, maxItems: 3, items: { $ref: '#/components/schemas/PermissionResponse' } } } }
        '401': { description: 会话无效, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: 无权访问应用, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: 应用不存在, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '503': { description: 权限服务暂时不可用, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    post:
      operationId: submitApplicationPermissions
      summary: 逐项提交当前应用的接口权限申请
      security: [{ ConsoleSession: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PermissionRequest' } } } }
      responses:
        '200':
          description: 权限申请已提交
          headers: { X-Request-ID: { required: true, schema: { type: string, minLength: 1 } } }
          content: { application/json: { schema: { type: array, minItems: 1, maxItems: 3, items: { $ref: '#/components/schemas/PermissionResponse' } } } }
        '400': { description: 请求无效, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: 会话无效, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: 无权访问应用或 CSRF 失败, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: 权限已待审或已通过, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '503': { description: 权限服务暂时不可用, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /registration-applications/csrf:
    get:
      operationId: getRegistrationCsrfToken
      summary: 获取匿名注册所需的 CSRF 令牌
      security: []
      responses:
        '200':
          description: CSRF token initialized in XSRF-TOKEN cookie
          content:
            application/json:
              schema:
                type: object
                required: [headerName, token]
                properties:
                  headerName: { type: string, const: X-XSRF-TOKEN }
                  token: { type: string }
        '403':
          description: 请求被安全策略拒绝
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /registration-applications:
    post:
      operationId: submitRegistrationApplication
      summary: 提交企业入驻申请
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegistrationApplicationRequest' }
      responses:
        '201':
          description: 申请已提交
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegistrationApplicationResponse' }
              example:
                applicationId: 0d65fa1b-ef36-46a4-8b1c-24585f867989
                status: PENDING_REVIEW
                submittedAt: '2026-08-14T09:00:00Z'
                reviewRole: 商务专员
                supportChannels: [企业微信, 邮件]
                nextAction: 等待商务专员审核，审核前无法使用平台功能
                requestId: req_example
        '400':
          description: 字段校验失败
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: CSRF 或安全策略拒绝请求
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: 账号已存在
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '413':
          description: 请求体超过 16 KiB
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '415':
          description: 请求媒体类型或字符集不受支持
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429':
          description: 匿名注册提交超过每 IP 每分钟 10 次
          headers:
            Retry-After:
              schema: { type: integer, minimum: 1 }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '500':
          description: 注册处理失败
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '503':
          description: 关键注册保护暂时不可用
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
components:
  schemas:
    Health: { $ref: './components/schemas.yaml#/components/schemas/Health' }
    Error: { $ref: './components/errors.yaml#/components/schemas/Error' }
    CsrfToken:
      type: object
      additionalProperties: false
      required: [headerName, token]
      properties:
        headerName: { type: string, const: X-XSRF-TOKEN }
        token: { type: string, minLength: 1 }
    LoginRequest:
      type: object
      additionalProperties: false
      required: [login, password]
      properties:
        login: { type: string, minLength: 4, maxLength: 64 }
        password: { type: string, minLength: 12, maxLength: 128, writeOnly: true }
    SessionContext:
      type: object
      additionalProperties: false
      required: [accountId, onboardingStatus, landingPath, requestId]
      properties:
        accountId: { type: string, format: uuid }
        onboardingStatus: { $ref: '#/components/schemas/OnboardingStatus' }
        landingPath: { type: string, enum: [/onboarding/status, /dashboard] }
        requestId: { type: string, minLength: 1 }
    OnboardingStatus: { type: string, enum: [PENDING_REVIEW, REJECTED, APPROVED] }
    OnboardingStatusResponse:
      type: object
      additionalProperties: false
      required: [status, submittedAt, updatedAt, rejectionReason, reviewRole, supportChannels, nextAction, requestId]
      properties:
        status: { $ref: '#/components/schemas/OnboardingStatus' }
        submittedAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        rejectionReason: { type: [string, 'null'], maxLength: 500 }
        reviewRole: { type: string, const: 商务专员 }
        supportChannels:
          type: array
          items: { type: string, enum: [企业微信, 邮件] }
        nextAction: { type: string, minLength: 1 }
        requestId: { type: string, minLength: 1 }
    ApplicationRequest:
      type: object
      additionalProperties: false
      required: [name, purpose]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        purpose: { type: string, minLength: 1, maxLength: 500 }
    ApplicationResponse:
      type: object
      additionalProperties: false
      required: [applicationId, name, purpose, appId, environment, status, createdAt, updatedAt, requestId]
      properties:
        applicationId: { type: string, format: uuid }
        name: { type: string }
        purpose: { type: string }
        appId: { type: string, minLength: 20 }
        environment: { type: string, const: SANDBOX }
        status: { type: string, const: ACTIVE }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        requestId: { type: string, minLength: 1 }
    ApplicationCreatedResponse:
      type: object
      additionalProperties: false
      required: [applicationId, name, purpose, appId, appSecret, environment, status, secretShownOnce, createdAt, updatedAt, requestId]
      properties:
        applicationId: { type: string, format: uuid }
        name: { type: string }
        purpose: { type: string }
        appId: { type: string, minLength: 20 }
        appSecret: { type: string, minLength: 43, description: 仅在本次创建响应返回一次 }
        environment: { type: string, const: SANDBOX }
        status: { type: string, const: ACTIVE }
        secretShownOnce: { type: boolean, const: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        requestId: { type: string, minLength: 1 }
    PermissionCode: { type: string, enum: [CUSTOMER_BASE_READ, ORDER_LIST_READ, ORDER_DETAIL_READ] }
    PermissionStatus: { type: string, enum: [NOT_APPLIED, PENDING_REVIEW, APPROVED, REJECTED] }
    PermissionRequest:
      type: object
      additionalProperties: false
      required: [permissions, reason]
      properties:
        permissions: { type: array, minItems: 1, maxItems: 3, uniqueItems: true, items: { $ref: '#/components/schemas/PermissionCode' } }
        reason: { type: string, minLength: 1, maxLength: 500 }
    PermissionResponse:
      type: object
      additionalProperties: false
      required: [code, name, purpose, dataScope, sensitiveNotice, status, submittedAt, updatedAt, rejectionReason, requestId]
      properties:
        code: { $ref: '#/components/schemas/PermissionCode' }
        name: { type: string, minLength: 1 }
        purpose: { type: string, minLength: 1 }
        dataScope: { type: string, minLength: 1 }
        sensitiveNotice: { type: string, minLength: 1 }
        status: { $ref: '#/components/schemas/PermissionStatus' }
        submittedAt: { type: [string, 'null'], format: date-time }
        updatedAt: { type: [string, 'null'], format: date-time }
        rejectionReason: { type: [string, 'null'], maxLength: 500 }
        requestId: { type: string, minLength: 1 }
    RegistrationApplicationRequest:
      type: object
      additionalProperties: false
      required: [enterpriseName, contactName, contactMobile, username, password]
      properties:
        enterpriseName: { type: string, minLength: 2, maxLength: 200 }
        contactName: { type: string, minLength: 2, maxLength: 100 }
        contactMobile:
          type: string
          description: 支持数字间的空格或连字符，服务端规范化后校验中国大陆手机号
          pattern: '^[\s\u00A0\u3000]*1[3-9](?:[\s\u00A0\u3000-]*[0-9]){9}[\s\u00A0\u3000]*$'
        username: { type: string, minLength: 4, maxLength: 64, pattern: '^[A-Za-z0-9._-]+$' }
        password: { type: string, minLength: 12, maxLength: 128, writeOnly: true }
    RegistrationApplicationResponse:
      type: object
      additionalProperties: false
      required: [applicationId, status, submittedAt, reviewRole, supportChannels, nextAction, requestId]
      properties:
        applicationId: { type: string, format: uuid }
        status: { type: string, enum: [PENDING_REVIEW] }
        submittedAt: { type: string, format: date-time }
        reviewRole: { type: string, const: 商务专员 }
        supportChannels:
          type: array
          items: { type: string, enum: [企业微信, 邮件] }
        nextAction: { type: string }
        requestId: { type: string }
  securitySchemes:
    ConsoleSession: { $ref: './components/security.yaml#/components/securitySchemes/ConsoleSession' }